Parse time as float.

+++ app/devices/vendors/shelly/gen_1/http/shelly_2p5.py
@@ -46,7 +46,7 @@ class Shelly2p5(ShellyHttpBase):
 
         # Set timer. (Default value is 1 second.)
         update_period = self._get_option("update_period", 1)
-        update_period = int(update_period)
+        update_period = float(update_period)
         self.__update_period = update_period
         self.__timer = Timer(self.__update_period)
         self.__timer.set_callback(self.__timer_cb)
